Refactored classes into functional folders
