+++ src/LibToyBot/Reporting/IPositionReporter.cs
@@ -1,0 +1,7 @@
+﻿namespace LibToyBot.Reporting
+{
+    public interface IPositionReporter
+    {
+        PositionReport Report();
+    }
+}
