+++ src/LibToyBot/Spatial/BoundsEvaluator.cs
