+++ src/LibToyBot/Spatial/IBoundsEvaluator.cs
