+++ src/LibToyBot/Spatial/IPositionTracker.cs
