+++ src/LibToyBot/Spatial/PositionTracker.cs
